seperate tests to suites
