+++ test/e2e/nftables_test.go
@@ -1,0 +1,90 @@
+package e2e
+
+import (
+	"log"
+	"strings"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	"golang.org/x/sync/errgroup"
+
+	"github.com/openshift-kni/commatrix/pkg/consts"
+
+	"github.com/openshift-kni/commatrix/pkg/types"
+	"github.com/openshift-kni/commatrix/test/pkg/firewall"
+	node "github.com/openshift-kni/commatrix/test/pkg/node"
+)
+
+var (
+	workerNodeRole = "worker"
+	tableName      = "table inet openshift_filter"
+	chainName      = "chain OPENSHIFT"
+)
+
+var _ = Describe("Nftables", func() {
+	It("should apply firewall by blocking all ports except the ones OCP is listening on", func() {
+		masterMat, workerMat := commatrix.SeparateMatrixByRole()
+		var workerNFT []byte
+
+		By("Creating NFT output for each role")
+		masterNFT, err := masterMat.ToNFTables()
+		Expect(err).NotTo(HaveOccurred())
+		if !isSNO {
+			workerNFT, err = workerMat.ToNFTables()
+			Expect(err).NotTo(HaveOccurred())
+		}
+
+		g := new(errgroup.Group)
+		for _, node := range nodeList.Items {
+			nodeName := node.Name
+			nodeRole, err := types.GetNodeRole(&node)
+			Expect(err).ToNot(HaveOccurred())
+			g.Go(func() error {
+				nftTable := masterNFT
+				if nodeRole == workerNodeRole {
+					nftTable = workerNFT
+				}
+
+				By("Applying firewall on node: " + nodeName)
+				err := firewall.ApplyRulesToNode(nftTable, nodeName, testNS, artifactsDir, utilsHelpers)
+				if err != nil {
+					return err
+				}
+				return nil
+			})
+		}
+
+		// Wait for all goroutines to finish
+		err = g.Wait()
+		Expect(err).ToNot(HaveOccurred())
+		nodeName := nodeList.Items[0].Name
+
+		By("Rebooting first node: " + nodeName + "and waiting for disconnect")
+
+		err = node.SoftRebootNodeAndWaitForDisconnect(utilsHelpers, cs, nodeName, testNS)
+		Expect(err).ToNot(HaveOccurred())
+
+		By("Waiting for node to be ready")
+		node.WaitForNodeReady(nodeName, cs)
+
+		debugPod, err := utilsHelpers.CreatePodOnNode(nodeName, testNS, consts.DefaultDebugPodImage)
+		Expect(err).ToNot(HaveOccurred())
+
+		defer func() {
+			err := utilsHelpers.DeletePod(debugPod)
+			Expect(err).ToNot(HaveOccurred())
+		}()
+
+		By("Listing nftables rules after reboot")
+		output, err := firewall.NftListAndWriteToFile(debugPod, utilsHelpers, artifactsDir, "nftables-after-reboot-"+nodeName)
+		Expect(err).ToNot(HaveOccurred())
+
+		By("Checking if nftables contain the chain OPENSHIFT after reboot")
+		if strings.Contains(string(output), tableName) &&
+			strings.Contains(string(output), chainName) {
+			log.Println("OPENSHIFT chain found in nftables.")
+		} else {
+			Fail("OPENSHIFT chain not found in nftables")
+		}
+	})
+})
