+++ test/e2e/validation_test.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 
-	"log"
 	"net/http"
 	"os"
 	"path/filepath"
@@ -15,38 +14,18 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/sync/errgroup"
 
 	configv1 "github.com/openshift/api/config/v1"
 	configv1client "github.com/openshift/client-go/config/clientset/versioned/typed/config/v1"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	clientOptions "sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/openshift-kni/commatrix/pkg/client"
 	commatrixcreator "github.com/openshift-kni/commatrix/pkg/commatrix-creator"
-	"github.com/openshift-kni/commatrix/pkg/consts"
-	"github.com/openshift-kni/commatrix/pkg/endpointslices"
 
 	listeningsockets "github.com/openshift-kni/commatrix/pkg/listening-sockets"
 	matrixdiff "github.com/openshift-kni/commatrix/pkg/matrix-diff"
 	"github.com/openshift-kni/commatrix/pkg/types"
-	"github.com/openshift-kni/commatrix/pkg/utils"
-	"github.com/openshift-kni/commatrix/test/pkg/firewall"
-	node "github.com/openshift-kni/commatrix/test/pkg/node"
-)
-
-var (
-	cs           *client.ClientSet
-	commatrix    *types.ComMatrix
-	isSNO        bool
-	isBM         bool
-	deployment   types.Deployment
-	infra        types.Env
-	utilsHelpers utils.UtilsInterface
-	epExporter   *endpointslices.EndpointSlicesExporter
-	nodeList     *corev1.NodeList
-	artifactsDir string
 )
 
 var (
@@ -85,73 +64,11 @@ const (
 )
 
 const (
-	workerNodeRole     = "worker"
-	tableName          = "table inet openshift_filter"
-	chainName          = "chain OPENSHIFT"
-	testNS             = "openshift-commatrix-test"
 	serviceNodePortMin = 30000
 	serviceNodePortMax = 32767
 )
 
-var _ = BeforeSuite(func() {
-	By("Creating output folder")
-	artifactsDir = os.Getenv("ARTIFACT_DIR")
-	if artifactsDir == "" {
-		log.Println("env var ARTIFACT_DIR is not set, using default value")
-	}
-	artifactsDir = filepath.Join(artifactsDir, "commatrix-e2e")
-
-	err := os.MkdirAll(artifactsDir, 0755)
-	Expect(err).NotTo(HaveOccurred())
-
-	By("Creating the clients for the Generating step")
-	cs, err = client.New()
-	Expect(err).NotTo(HaveOccurred())
-
-	utilsHelpers = utils.New(cs)
-
-	deployment = types.Standard
-	isSNO, err = utilsHelpers.IsSNOCluster()
-	Expect(err).NotTo(HaveOccurred())
-
-	if isSNO {
-		deployment = types.SNO
-	}
-
-	infra = types.Cloud
-	isBM, err = utilsHelpers.IsBMInfra()
-	Expect(err).NotTo(HaveOccurred())
-
-	if isBM {
-		infra = types.Baremetal
-	}
-
-	epExporter, err = endpointslices.New(cs)
-	Expect(err).ToNot(HaveOccurred())
-
-	By("Generating comMatrix")
-	commMatrixCreator, err := commatrixcreator.New(epExporter, "", "", infra, deployment)
-	Expect(err).NotTo(HaveOccurred())
-
-	commatrix, err = commMatrixCreator.CreateEndpointMatrix()
-	Expect(err).NotTo(HaveOccurred())
-
-	By("Creating test namespace")
-	err = utilsHelpers.CreateNamespace(testNS)
-	Expect(err).ToNot(HaveOccurred())
-
-	nodeList = &corev1.NodeList{}
-	err = cs.List(context.TODO(), nodeList)
-	Expect(err).ToNot(HaveOccurred())
-})
-
-var _ = AfterSuite(func() {
-	By("Deleting Namespace")
-	err := utilsHelpers.DeleteNamespace(testNS)
-	Expect(err).ToNot(HaveOccurred())
-})
-
-var _ = Describe("commatrix", func() {
+var _ = Describe("Validation", func() {
 	It("generated communication matrix should be equal to documented communication matrix", func() {
 		By("get cluster's version and check if it's suitable for test")
 		clusterVersion, err := getClusterVersion(cs)
@@ -173,7 +90,7 @@ var _ = Describe("commatrix", func() {
 		resp, err := http.Get(strings.Replace(docCommatrixBaseURL, "VERSION", clusterVersion, 1))
 		Expect(err).ToNot(HaveOccurred())
 		defer resp.Body.Close()
-		// if response status code equals to "status not found", compare generated commatrix to the master documented commatrix
+		// if response status code equals to "status not found", compare generated commatrix to the main documented commatrix
 		if resp.StatusCode == http.StatusNotFound {
 			resp, err = http.Get(strings.Replace(docCommatrixBaseURL, "enterprise-VERSION", "main", 1))
 			Expect(err).ToNot(HaveOccurred())
@@ -226,13 +143,13 @@ var _ = Describe("commatrix", func() {
 		// Get ports that are in the documented commatrix but not in the generated commatrix.
 		notUsedPortsMat := diff.GenerateUniqueSecondary()
 		if len(notUsedPortsMat.Matrix) > 0 {
-			logrus.Warningf("the following ports are documented but are not used: \n %s", notUsedPortsMat)
+			logrus.Warningf("the following ports are documented but are not used:\n%s", notUsedPortsMat)
 		}
 
 		// Get ports that are in the generated commatrix but not in the documented commatrix.
 		missingPortsMat := diff.GenerateUniquePrimary()
 		if len(missingPortsMat.Matrix) > 0 {
-			err := fmt.Errorf("the following ports are used but are not documented: \n %s", missingPortsMat)
+			err := fmt.Errorf("the following ports are used but are not documented:\n%s", missingPortsMat)
 			Expect(err).ToNot(HaveOccurred())
 		}
 	})
@@ -274,72 +191,6 @@ var _ = Describe("commatrix", func() {
 		if len(missingEPSMat.Matrix) > 0 {
 			err := fmt.Errorf("the following ports are used but don't have an endpointslice: \n %s", missingEPSMat)
 			Expect(err).ToNot(HaveOccurred())
-		}
-	})
-
-	It("should apply firewall by blocking all ports except the ones OCP is listening on", func() {
-		masterMat, workerMat := commatrix.SeparateMatrixByRole()
-		var workerNFT []byte
-
-		By("Creating NFT output for each role")
-		masterNFT, err := masterMat.ToNFTables()
-		Expect(err).NotTo(HaveOccurred())
-		if !isSNO {
-			workerNFT, err = workerMat.ToNFTables()
-			Expect(err).NotTo(HaveOccurred())
-		}
-
-		g := new(errgroup.Group)
-		for _, node := range nodeList.Items {
-			nodeName := node.Name
-			nodeRole, err := types.GetNodeRole(&node)
-			Expect(err).ToNot(HaveOccurred())
-			g.Go(func() error {
-				nftTable := masterNFT
-				if nodeRole == workerNodeRole {
-					nftTable = workerNFT
-				}
-
-				By("Applying firewall on node: " + nodeName)
-				err := firewall.ApplyRulesToNode(nftTable, nodeName, testNS, artifactsDir, utilsHelpers)
-				if err != nil {
-					return err
-				}
-				return nil
-			})
-		}
-
-		// Wait for all goroutines to finish
-		err = g.Wait()
-		Expect(err).ToNot(HaveOccurred())
-		nodeName := nodeList.Items[0].Name
-
-		By("Rebooting first node: " + nodeName + "and waiting for disconnect")
-
-		err = node.SoftRebootNodeAndWaitForDisconnect(utilsHelpers, cs, nodeName, testNS)
-		Expect(err).ToNot(HaveOccurred())
-
-		By("Waiting for node to be ready")
-		node.WaitForNodeReady(nodeName, cs)
-
-		debugPod, err := utilsHelpers.CreatePodOnNode(nodeName, testNS, consts.DefaultDebugPodImage)
-		Expect(err).ToNot(HaveOccurred())
-
-		defer func() {
-			err := utilsHelpers.DeletePod(debugPod)
-			Expect(err).ToNot(HaveOccurred())
-		}()
-
-		By("Listing nftables rules after reboot")
-		output, err := firewall.NftListAndWriteToFile(debugPod, utilsHelpers, artifactsDir, "nftables-after-reboot-"+nodeName)
-		Expect(err).ToNot(HaveOccurred())
-
-		By("Checking if nftables contain the chain OPENSHIFT after reboot")
-		if strings.Contains(string(output), tableName) &&
-			strings.Contains(string(output), chainName) {
-			log.Println("OPENSHIFT chain found in nftables.")
-		} else {
-			Fail("OPENSHIFT chain not found in nftables")
 		}
 	})
 })
